Define custom scrollbar for windows
As discussed in comments, the windows scrollbar is very intrusive, so
this commit defines an OSX like scrollbar. However,
::-webkit-scrollbar-* is very finicky, and won't work if defined within
another block like body { ::-webkit-scrollbar-* {} }, which breaks a few
things like nightmode.

Instead, this commit will use javascript to inject a style into <head>,
only if on windows, and using the nightmode appropriate color