Add initial ReasonReact build setup
Adds Reason and ReasonReact as dependencies, and creates a test main.re
that gets compiled as part of the build (now moved to build.sh) before
being included in index.html.

The build is less than ideal, because bucklescript insists on using
lib/bs as the cache dir, and leaving *.bs.js compiled javascript files
next to source files, but I'll live with this for now to keep the build
script simple.