Move modal state up to Main, modal is child of main browser
As part of making Modal a child component of the main browser, I needed
to bubble up the state to a global level, because things like the menu and
each image's onclick depend or send actions to affect the modal state.

This involves some odd code organization, as I'd like to keep the modal
logic within the modal module as much as possible. However, I ended up
pulling the reducer into Main, because the return type needs to conform
to the Main reducer, and I could not pull in Main types in modal without
a circular reference.