Make ImageClick a reducer action
Because imageOnClick depended on Main's mode, on every mode change, all
the images would have to be re-rendered (because of the prop change).
This led to some lag on larger directories.

This makes ImageClick a single action, which is probably a paradigm to
follow. However, without a good way of composing reducers, this is super
cumbersome around setting the modal.