Modify zoom algorithm so ZoomIn/ZoomOut always respond
Previously zooming could not cause a visual change, e.g. if the window
is 1200px and colw goes from 500-600, there would still be two columns
of 600px.

This commit changes the functionality so that zooming always results in
a column count change, by increasing/decreasing the number of columns
based on the current window width.