Only include minimum bs-platform, reason-react in app
Adds a step in the build phase to
- Extract local copy of bs-platform/lib/js and reason-react into build/
- Replace references in compiled *.bs.js and reason-react *.js files to
point to local modules

The bs-platform package does not expose the runtime libraries as a
separate module, so the whole package must be included as part of the
build.

I believe on Windows this is causing some wonkiness as startup times are
reaching 30-60seconds, with a lot of disk activity. It appears like
there's a large amount (300-500mb) written and deleted on every app
start to ~/AppData/Local/Temp/... containing only the bs-platform
package in an asar.unpacked

This is an attempt at a workaround, by manually limiting the build to
only the necessary javascript files