Add platform class to body, windows specific draggable regions
Frameless electron windows aren't draggable. This is easily solved on
MacOS by adding `-webkit-app-region: drag` to anywhere that should drag,
and since text isn't selectable in Isolate, we can uniformly add it
everywhere.

However, on Windows, draggable elements do not register mouse events.
This commit adds a windows only top and bottom draggable region to deal
with this, so the bottom/top of the window can still be dragged, while
the rest is interactive. Not ideal, but some what of a workaround.